+++ byproxy/__init__.py
@@ -5,11 +5,4 @@ ByProxy is a simple package contains simple tools for proxy management and usage
 
 """
 from byproxy.proxy_maker import ProxyMaker
-from byproxy.proxy_checker import ProxyChecker
-
-__version__ = '0.1.0'
-__author__ = 'vimevim <vimevim@gmail.com>'
-__all__ = [
-    'ProxyMaker',
-    'ProxyChecker',
-    ]
+from byproxy.proxy_checker import async_api, sync_api