+++ CHANGELOG.md
@@ -5,11 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 0.1.4 - 2022-12-31
+
+### Changed
+
+* ProxyChecker file structure is changed. Now it's more moduler.
+
+### Added
+
+* async_api for ProxyChecker class.
+
+### Changed
+
+* Improved documentation of ProxyMaker class and added some expected outputs.
+* Improved ProxyMaker make_proxy method, now we can pick a library for making proxies in the format we want.
+
 ## Version 0.1.3.3 - 2022-12-26
 
 ### Internal
 
 * Updated ProxyChecker class methods, they don't return dict anymore, they make the request and return the response object.
+
 ## Version 0.1.3.2 - 2022-12-26
 
 ### Fixed
@@ -24,7 +40,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Split every conversion method in ProxyMake class into a single version of it and a list version of it.
+* Added new methods to ProxyMaker class:
+  * split_line
+  * line_to_dict
+
+### Internal
+
+* ProxyMaker class has few updates:
+  * split_lines uses split_line method.
+  * lines_to_dicts uses line_to_dict method.
 
 ## Version 0.1.3 - 2022-12-26